Generalize getting path from route and entity
