+++ node/middlewares/internals/utils.ts
@@ -1,3 +1,6 @@
+import { Apps } from '@vtex/api'
+import RouteParser from 'route-parser'
+
 export const tenMinutesFromNowMS = () =>
   `${new Date(Date.now() + 10 * 60 * 1000)}`
 export const STORE_LOCATOR = 'vtex.store@2.x'
@@ -24,3 +27,23 @@ export interface ContentTypeDefinition {
 }
 
 export type Routes = Record<string, ContentTypeDefinition>
+type PageTypes = typeof PAGE_TYPES[keyof typeof PAGE_TYPES]
+
+export const getPath = async (
+  type: PageTypes,
+  params: any,
+  apps: Apps
+): Promise<string> => {
+  const routesJSON = await apps.getAppJSON<Routes>(
+    STORE_LOCATOR,
+    ROUTES_JSON_PATH
+  )
+  const route = routesJSON[type]
+  const canonicalParser = new RouteParser(route.canonical)
+
+  const path = canonicalParser.reverse(params)
+  if (!path) {
+    throw new Error(`Parse error, params: ${params}, path: ${path}`)
+  }
+  return path
+}
